+++ bboss-elasticsearch-rest/src/main/java/org/frameworkset/elasticsearch/client/AbstractSlowDslCallback.java
@@ -1,4 +1,4 @@
-package org.frameworkset.elasticsearch.entity;
+package org.frameworkset.elasticsearch.client;
 /**
  * Copyright 2008 biaoping.yin
  * <p>
@@ -19,19 +19,11 @@ package org.frameworkset.elasticsearch.entity;
  * <p>Description: </p>
  * <p></p>
  * <p>Copyright (c) 2018</p>
- * @Date 2019/11/13 14:08
+ * @Date 2019/11/13 12:52
  * @author biaoping.yin
  * @version 1.0
  */
-public class SlowDsl extends LogDsl{
+public abstract class AbstractSlowDslCallback implements SlowDslCallback{
 
-	private  int slowDslThreshold;
 
-	public int getSlowDslThreshold() {
-		return slowDslThreshold;
-	}
-
-	public void setSlowDslThreshold(int slowDslThreshold) {
-		this.slowDslThreshold = slowDslThreshold;
-	}
 }