+++ bboss-elasticsearch-rest/src/main/java/org/frameworkset/elasticsearch/client/DefaultSlowDslCallback.java
@@ -1,4 +1,4 @@
-package org.frameworkset.elasticsearch.entity;
+package org.frameworkset.elasticsearch.client;
 /**
  * Copyright 2008 biaoping.yin
  * <p>
@@ -15,23 +15,28 @@ package org.frameworkset.elasticsearch.entity;
  * limitations under the License.
  */
 
+import org.frameworkset.elasticsearch.entity.SlowDsl;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 /**
  * <p>Description: </p>
  * <p></p>
  * <p>Copyright (c) 2018</p>
- * @Date 2019/11/13 14:08
+ * @Date 2019/11/13 12:52
  * @author biaoping.yin
  * @version 1.0
  */
-public class SlowDsl extends LogDsl{
+public class DefaultSlowDslCallback extends AbstractSlowDslCallback{
 
-	private  int slowDslThreshold;
+	private static final Logger logger = LoggerFactory.getLogger(DefaultSlowDslCallback.class);
+	@Override
+	public void slowDslHandle(SlowDsl slowDsl){
+		if(logger.isWarnEnabled()) {
+			logger.warn("Slow request[{}] action[{}] took time:{} ms > slowDslThreshold[{} ms], use DSL[{}]",
+					slowDsl.getUrl(),slowDsl.getAction(), slowDsl.getTime(), slowDsl.getSlowDslThreshold(),  RestSearchExecutorUtil.chunkEntity(slowDsl.getDsl()));
 
-	public int getSlowDslThreshold() {
-		return slowDslThreshold;
+		}
 	}
 
-	public void setSlowDslThreshold(int slowDslThreshold) {
-		this.slowDslThreshold = slowDslThreshold;
-	}
 }